+++ src/_/install.ts
@@ -1,6 +1,6 @@
 import type { Options, Environment } from "./types";
 import { createElement } from "./dom";
-import { isBrowser, isReactNative, logWarning, throwError } from "./utils";
+import { isBrowser, isReactNative, logWarning, noop, throwError } from "./utils";
 import { rValidEnvironment } from "./validate";
 import { VERSION } from "./const";
 
@@ -12,16 +12,18 @@ const defaultClass = `twic`;
 export const config: {
     debug: boolean,
     class: string,
-    env: Environment,
     domain: string,
+    env: Environment,
+    handleShadowDom: ( item?: Element ) => void,
     maxDPR: number;
     path: string,
     step: number,
 } = {
     "debug": false,
     "class": defaultClass,
-    "env": `production`,
     "domain": undefined,
+    "env": `production`,
+    "handleShadowDom": noop,
     "maxDPR": undefined,
     "path": ``,
     "step": undefined,
@@ -43,6 +45,43 @@ export const getDataAttributeName = ( baseName: string ): string => `data-${ con
 const rInvalidPath = /\?|^\/*$/;
 const rValidDomain = /(^https?:\/\/[^/]+)\/*$/;
 
+const handleShadowDomFactory = ( attributeName: string ) => {
+    const marked = new WeakSet();
+    return ( item: Element ): void => {
+        while ( item && !marked.has( item ) ) {
+            marked.add( item );
+            const { parentNode } = item;
+            const isHost = !parentNode && ( item instanceof ShadowRoot );
+            if ( isHost ) {
+                if ( ( item as unknown as ShadowRoot ).mode === `closed` ) {
+                    throw new Error( `cannot use TwicPics components in closed ShadowRoot` );
+                }
+                createElement( {
+                    "element": item as unknown as ShadowRoot,
+                    "value": {
+                        "elementName": `style`,
+                        "value": `*STYLE*${ configBasedStyle() }`,
+                    },
+                } );
+                // eslint-disable-next-line no-param-reassign
+                item = ( item as unknown as ShadowRoot ).host;
+                if ( item ) {
+                    item.setAttribute( attributeName, `` );
+                }
+            } else {
+                // eslint-disable-next-line no-param-reassign
+                item = ( parentNode as Element );
+            }
+        }
+    };
+};
+
+const parametersMap = [
+    [ `anticipation`, `anticipation` ],
+    [ `class`, `class` ],
+    [ `maxDPR`, `max-dpr` ],
+    [ `step`, `step` ],
+];
 export default ( options: Options ): void => {
 
     if ( !options ) {
@@ -50,7 +89,7 @@ export default ( options: Options ): void => {
     }
 
     const hasPreviousInstall = config && config.domain;
-    const { domain, "class": _class, env, path } = options;
+    const { domain, "class": _class, env, handleShadowDom, path } = options;
 
     if ( !domain || !rValidDomain.test( domain ) ) {
         throwError( `install domain "${ domain }" is invalid` );
@@ -68,6 +107,10 @@ export default ( options: Options ): void => {
     config.domain = domain.replace( rValidDomain, `$1` );
     config.env = env;
     config.path = path ? path.replace( /^\/*(.+?)\/*$/, `$1/` ) : ``;
+    config.handleShadowDom =
+        ( handleShadowDom && isBrowser && !isReactNative ) ?
+            handleShadowDomFactory( getDataAttributeName( `component` ) ) :
+            noop;
 
     if ( isReactNative ) {
         const { debug, maxDPR, step } = options;
@@ -82,41 +125,41 @@ export default ( options: Options ): void => {
             logWarning( `install function called multiple times` );
             return;
         }
-
         const parts = [ `${ config.domain }/?${ VERSION }` ];
-        Object.entries( options ).forEach( ( [ key, value ] ) => {
-            if ( value != null ) {
-                let actualKey = key;
-                if ( key === `maxDPR` ) {
-                    actualKey = `max-dpr`;
-                }
-                if ( ( key !== `domain` ) && ( key !== `path` ) && ( key !== `mode` ) ) {
-                    parts.push( `${ actualKey }=${ value }` );
+
+        parametersMap.forEach( p => {
+            const [ key, actualKey ] = p;
+            if ( options.hasOwnProperty( key ) ) {
+                const value = options[ key as keyof Options ];
+                if ( value ) {
+                    parts.push( `${ actualKey }=${ options[ key as keyof Options ] }` );
                 }
             }
         } );
 
-        createElement( [
-            document.head,
-            0,
-            [
-                [
-                    `link`,
-                    {
+        createElement( {
+            "element": document.head,
+            "value": [
+                {
+                    "attributes": {
                         "rel": `preconnect`,
                         "href": domain,
                     },
-                ],
-                [
-                    `script`,
-                    {
+                    "elementName": `link`,
+                },
+                {
+                    "attributes": {
                         "async": ``,
                         "defer": ``,
                         "src": parts.join( `&` ),
                     },
-                ],
-                [ `style`, 0, configBasedStyle() ],
+                    "elementName": `script`,
+                },
+                {
+                    "value": configBasedStyle(),
+                    "elementName": `style`,
+                },
             ],
-        ] );
+        } );
     }
 };